Added support for the default namespace in the VB projects.
The part of infrastructure which is shared across different languages for project file generation factored out. The reason to not include that in the Language class is to keep it clean from the implementation details of the IDE for which project file will be generated. Technically possible that project files would be generated not only for VS.